Create README.md
This is a Social Media Integration application which was created using the java language, it is an application that consists of four social media platforms namely Facebook, Instagram, Twitter and Telegram. I created this application mainly to use as a dual space for four platforms to maintain multiple accounts at the same time.